requirements.txt
google-cloud-bigquery == 3.9.0
pandas == 1.3.5
db-dtypes == 1.1.1
google-cloud-pubsub == 2.16.0